Some changes to support older Delphi versions

+++ Demos/01.HelloWorld/Demo.Entities.pas
@@ -61,7 +61,7 @@ type
   end;
 
   TArrayParam = TArray<TRecordParam>;
-  TArrayInt = TArray<Integer>;
+  TArrayInt = TArray<Int64>;
 
   TSimpleParam = class(TObject)
   private
